% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{calc_adj_frag}
\alias{calc_adj_frag}
\title{Calculate adjacency matrices and fragility from voltage readings}
\usage{
calc_adj_frag(readings, t_window, t_step, lambda = NULL)
}
\arguments{
\item{readings}{Numeric. A matrix of voltage readings,
with time points as rows and electrodes as columns}

\item{t_window}{Integer. The number of time points to use in each window}

\item{t_step}{Integer.The number of time points to move the window each time}

\item{lambda}{Numeric. The lambda value to use in the ridge regression.
If NULL, the lambda will be chosen automatically}
}
\value{
A list containing the normalized readings,
adjacency matrices, fragility, and R^2 values
}
\description{
A very long long description to give the details of the function
}
\examples{
data <- matrix(rnorm(100), nrow = 10)
t_window <- 10
t_step <- 5
lambda <- 0.1
calc_adj_frag(readings = data, t_window = t_window, t_step = t_step, lambda = lambda)

}
