% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_subject}
\alias{check_subject}
\title{check_subject}
\usage{
check_subject(subject_code, subject_dir, trials)
}
\arguments{
\item{subject_code}{}

\item{subject_dir}{}

\item{trials}{}
}
\description{
check_subject
}
