% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{dev_Fragility}
\alias{dev_Fragility}
\title{Function to load all development functions within an environment}
\usage{
dev_Fragility(expose_functions = FALSE, reload = TRUE)
}
\arguments{
\item{expose_functions}{logical indicating whether to expose all development 
functions to the global environment}

\item{reload}{logical, do you want to fast-reload the package before load the 
functions?}
}
\description{
Function to load all development functions within an environment
}
