% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragility_functions.R
\name{draw_f_map_table}
\alias{draw_f_map_table}
\title{draw_f_map_table}
\usage{
draw_f_map_table(
  tnum,
  adj_info,
  f_path,
  subject_code,
  requested_electrodes,
  sort_fmap,
  check,
  f_list_length
)
}
\arguments{
\item{tnum}{An integer vector specifying which trials are selected. If
multiple trials are selected, their fragility maps/values will be averaged.}

\item{adj_info}{The adj_info file generated by generate_adj_array.}

\item{f_path}{Pathname to subject's fragility files, with the trial number
left empty. Will generally be something similar to
'rave_data/data_dir/ProjectName/SubjectCode/rave/module_data/SubjectCode_f_info_trial_'}

\item{subject_code}{Character object specifying subject's code.}

\item{requested_electrodes}{Integer vector specifying which electrodes to
display.}

\item{sort_fmap}{Specifies whether to sort map by electrode number or by
fragility. Will be either "Electrode" or "Fragility".}

\item{check}{Check list generated by check_subject.}

\item{f_list_length}{Integer specifying how many values should be included on
the most/least fragile list.}
}
\value{

}
\description{
Generates the parameters needed for the 3D brain viewer, fragility map,
most/least fragility table, and selected trials outputs.
}
\examples{
outputs <- draw_f_map_table(
    tnum = c(1,2,3), 
    adj_info = adj_info, 
    f_path = 'rave_data/data_dir/OnsetZone/PT01/rave/module_data/PT01_f_info_trial_', 
    subject_code = 'PT01', 
    requested_electrodes = c(1:24,26:36,42:43,46:54,56:70,72:95), 
    sort_fmap = 'Electrodes', 
    check = check, 
    f_list_length = 10
)
}
