% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{draw_heat_map}
\alias{draw_heat_map}
\title{draw_heat_map
this function makes an image using matrix mat}
\usage{
draw_heat_map(mat, x, y, zlim)
}
\arguments{
\item{mat}{a 2D matrix}

\item{x}{vector for x axis}

\item{y}{vector for y axis}

\item{zlim}{vector of length 2 specifying range of image}
}
\value{
heatmap image
}
\description{
draw_heat_map
this function makes an image using matrix mat
}
\examples{
draw_heat_map(mat, x, y)
}
