% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragility_functions.R
\name{find_adj_matrix}
\alias{find_adj_matrix}
\title{find_adj_matrix}
\usage{
find_adj_matrix(state_vectors, N, t_window, nlambda)
}
\arguments{
\item{state_vectors}{List of 2 state vectors x(t) and x(t+1) generated by
generate_state_vectors}

\item{N}{Number of electrodes}

\item{t_window}{Length of one time window}

\item{nlambda}{Number of lambdas to be generated by glmnet cross-validation}
}
\value{

}
\description{
Finds the adjacency matrix A for a given time window. For use within the
generate_adj_array function.
}
\examples{
A is the 3D adjacency array in the following example, where k is the timewindow being calculated.
A[,,k] <- find_adj_matrix(state_vectors, N = 85, t_window = 250, nlambda = 16)
}
