% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragility_functions.R
\name{find_fragility}
\alias{find_fragility}
\title{find_fragility
Finds the fragility value for a single electrode (node) in a single time window}
\usage{
find_fragility(node, A_k, N, limit)
}
\arguments{
\item{node}{The node that you're calculating the fragility for}

\item{N}{The number of electrodes}

\item{A}{The adjacency matrix that you're calculating the fragility for}
}
\value{

}
\description{
find_fragility
Finds the fragility value for a single electrode (node) in a single time window
}
