% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragility_functions.R
\name{generate_fragility_matrix}
\alias{generate_fragility_matrix}
\title{generate_fragility_matrix}
\usage{
generate_fragility_matrix(A, elec, lim = 0+1i)
}
\arguments{
\item{A}{NxNxJ adjacency array generated by generate_adj_array.}

\item{elec}{Character or integer vector containing all the included
electrodes.}

\item{lim}{Optional parameter setting the eigenvalue used for fragility
calculations. Must be an imaginary or complex number with a absolute value
greater than or equal to 1. Default lim is 1i.}
}
\value{
List containing the raw fragility data (lower number = more fragile),
  normalized fragility data from 0 to 1 (1 is most fragile), and average
  fragility data across all time windows.
}
\description{
This function generates the NxJ fragility matrix, where N is the number of
electrodes and J is the number of timewindows.
}
\examples{
f_info <- generate_fragility_matrix(
    A = adj_info$A, 
    elec = c(1:24,26:36,42:43,46:54,56:70,72:95)
)
}
