% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragility_functions.R
\name{generate_state_vectors}
\alias{generate_state_vectors}
\title{generate_state_vectors}
\usage{
generate_state_vectors(v, trial, t_window, t_start)
}
\arguments{
\item{t_start}{}
}
\value{

}
\description{
Generates state vectors x(t) and x(t+1) for use in finding the adjacency
matrix A in the equation x(t+1) = Ax(t).
}
