% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragility_functions.R
\name{load_fragility_patient}
\alias{load_fragility_patient}
\title{load_fragility_patient
this function loads a previously processed patient}
\usage{
load_fragility_patient(v, unit, halve = FALSE)
}
\arguments{
\item{halve}{Logical determining whether you want to halve the Hz value, e.g. from 2000Hz to 1000Hz (saves some computation time)}

\item{subject_code}{Subject code}

\item{block}{The epoch condition/trial that you're calculating for}

\item{elec}{Vector of electrodes you want included}
}
\value{
list containing the voltage trace matrix v, condition,
}
\description{
load_fragility_patient
this function loads a previously processed patient
}
