% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragility_functions.R
\name{process_fragility_patient}
\alias{process_fragility_patient}
\title{process_fragility_patient}
\usage{
process_fragility_patient(v, unit, srate, halve = FALSE)
}
\arguments{
\item{v}{3D voltage array from RAVE's module_tools$get_voltage() and
get_data() functions. First dimension is trial, second dimension is time,
and third dimension is electrode.}

\item{unit}{Character: 'uV', 'nV', or 'mV' specifying the units of the
voltage data.}

\item{halve}{Logical specifying whether the data should be halved or not.
This is for data that has a high sampling rate - for example, 2000 Hz data
would be processed much more efficiently if halved to 1000Hz, while still
retaining most of the important information.}
}
\value{
List containing the processed voltage trace matrix v as well as
  metadata about the PROCESSED data (trial numbers, unit, sampling rate) -
  NOT the original data!
}
\description{
This function processes voltage data in preparation for adjacency matrix and
fragility calculations. It converts the units of the data to uV if possible
and halves the frequency if necessary.
}
\examples{
voltage <- module_tools$get_voltage()
v <- voltage$get_data()
pt_info <- process_fragility_patient(v = v, unit = 'mV', halve = FALSE)
}
