% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rigde.r
\name{ridge}
\alias{ridge}
\title{Ridge Regression for Electrode Readings}
\usage{
ridge(xt, xtp1, lambda, intercept = FALSE, iw)
}
\arguments{
\item{xt}{matrix. What does row and column mean?}

\item{xtp1}{matrix. the electrode readings in the next time point}
}
\description{
A long long description.
}
\examples{
...

}
