% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_subject}
\alias{check_subject}
\title{check_subject}
\usage{
check_subject(subject_code, subject_dir, trials)
}
\arguments{
\item{subject_code}{Character specifying subject's code}

\item{subject_dir}{List containing sub-directories of subject's folder in
rave_data, from module_tools$get_subject_dirs()}

\item{trials}{Integer vector with a number for each trial}
}
\description{
check_subject
}
\examples{
check <- check_subject('KAA', module_tools$get_subject_dirs(), 1:8)
}
