% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragility_functions.R
\name{find_adj_matrix}
\alias{find_adj_matrix}
\title{find_adj_matrix}
\usage{
find_adj_matrix(state_vectors, N, t_window, nlambda, ncores)
}
\arguments{
\item{state_vectors}{}

\item{N}{}

\item{t_window}{}

\item{nlambda}{}

\item{ncores}{}
}
\value{

}
\description{
Finds the adjacency matrix A for a given time window
}
