% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragility_functions.R
\name{find_fragility}
\alias{find_fragility}
\title{find_fragility}
\usage{
find_fragility(node, A_k, N, limit)
}
\arguments{
\item{node}{Integer specifying the node whose fragility is being calculated}

\item{N}{Integer specifying number of electrodes}

\item{A}{The adjacency matrix of the given time window}
}
\value{

}
\description{
Finds the fragility value for a single electrode (node) in a
single time window
}
\examples{

}
