% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragility_functions.R
\name{generate_state_vectors}
\alias{generate_state_vectors}
\title{generate_state_vectors}
\usage{
generate_state_vectors(v, trial, t_window, t_start)
}
\arguments{
\item{v}{3D voltage matrix}

\item{trial}{trial number}

\item{t_window}{timewindow size}

\item{t_start}{start time of current timewindow}
}
\value{

}
\description{
Generates state vectors x(t) and x(t+1) for use in finding the adjacency
matrix A in the equation x(t+1) = Ax(t). For use within the
generate_adj_array function.
}
\examples{
state_vectors <- generate_state_vectors(v,trial = 1,t_window = 250,t_start = 1)
}
